review(server): constant for compare object
