+++ packages/twenty-server/src/engine/metadata-modules/flat-object-metadata/constants/flat-object-metadata-jsonb-properties.constant.ts
@@ -1,0 +1,5 @@
+import { FlatObjectMetadata } from "src/engine/metadata-modules/flat-object-metadata/types/flat-object-metadata.type";
+
+export const FLAT_OBJECT_METADATA_JSONB_PROPERTIES = [
+  'standardOverrides',
+] as const satisfies (keyof FlatObjectMetadata)[];
