+++ packages/twenty-server/src/engine/metadata-modules/flat-object-metadata/types/flat-object-metadata-jsonb-properties.type.ts
@@ -1,0 +1,4 @@
+import { FLAT_OBJECT_METADATA_JSONB_PROPERTIES } from "src/engine/metadata-modules/flat-object-metadata/constants/flat-object-metadata-jsonb-properties.constant";
+
+export type FlatObjectMetadataEntityJsonbProperties =
+  (typeof FLAT_OBJECT_METADATA_JSONB_PROPERTIES)[number];
