+++ packages/twenty-server/src/engine/metadata-modules/flat-object-metadata/types/flat-object-metadata-properties-to-compare.type.ts
@@ -1,0 +1,4 @@
+import { FLAT_OBJECT_METADATA_PROPERTIES_TO_COMPARE } from "src/engine/metadata-modules/flat-object-metadata/constants/flat-object-metadata-properties-to-compare.constant";
+
+export type FlatObjectMetadataPropertiesToCompare =
+  (typeof FLAT_OBJECT_METADATA_PROPERTIES_TO_COMPARE)[number];
