+++ packages/twenty-server/src/engine/metadata-modules/flat-object-metadata/utils/from-update-object-input-to-flat-object-metadata.util.ts
@@ -6,11 +6,9 @@ import {
 
 import { type FlatObjectMetadataMaps } from 'src/engine/metadata-modules/flat-object-metadata-maps/types/flat-object-metadata-maps.type';
 import { findFlatObjectMetadataInFlatObjectMetadataMaps } from 'src/engine/metadata-modules/flat-object-metadata-maps/utils/find-flat-object-metadata-in-flat-object-metadata-maps.util';
+import { FLAT_OBJECT_METADATA_PROPERTIES_TO_COMPARE } from 'src/engine/metadata-modules/flat-object-metadata/constants/flat-object-metadata-properties-to-compare.constant';
+import { FlatObjectMetadataPropertiesToCompare } from 'src/engine/metadata-modules/flat-object-metadata/types/flat-object-metadata-properties-to-compare.type';
 import { type FlatObjectMetadata } from 'src/engine/metadata-modules/flat-object-metadata/types/flat-object-metadata.type';
-import {
-  type FlatObjectMetadataPropertiesToCompare,
-  flatObjectMetadataPropertiesToCompare,
-} from 'src/engine/metadata-modules/flat-object-metadata/utils/compare-two-flat-object-metadata.util';
 import {
   type ObjectMetadataStandardOverridesProperties,
   objectMetadataStandardOverridesProperties,
@@ -28,7 +26,7 @@ type FromUpdateObjectInputToFlatObjectMetadataArgs = {
 };
 
 const objectMetadataEditableProperties =
-  flatObjectMetadataPropertiesToCompare.filter(
+  FLAT_OBJECT_METADATA_PROPERTIES_TO_COMPARE.filter(
     (
       property,
     ): property is Exclude<
